Start a completely new approach to the application
In the previous version of `cargo-valgrind` was relatively complicated.
There was a custom parsing of the cargo metadata, much knowledge regar-
ding the path to executable. It didn't work for custom targets, unit
tests were not supported, and many other restrictions.

This commit uses a completely different approach to the problem. Instead
of simulating a cargo frontend, cargo itself is used to invoke the pro-
gram, as cargo has all the knowledge. This application acts therefore
only as a proxy, that sets a specific environment variable and calls
`cargo` with all arguments the user passed.
This variable controls the program invoked by cargo when it tries to
start an executable. Currently it only calls directly `valgrind`, but
the old XML parsing etc. is not yet again in place.

Currently there is only the host architecture supported, but the program
will be extended to also support the `--target` flag.