Add a simple help page to `cargo valgrind`
Previously, `cargo valgrind --help` or `cargo help valgrind` did output
the help of `cargo` instead of the help of `cargo-valgrind`. This was
due to the program successfully registering itself as a runner, but the
normal `cargo --help` will be executed.
A pre-check for help flags is issued to prevent the old behavior.