Store the error output if valgrind failed
This improves the `Error::ValgrindFailure`, by capturing the standard
error of the `valgrind` binary. This change is rather simple.
While testing that change locally, it was unveiled, that cargo valgrind
will hang forever, if `valgrind` itself is called with wrong parameters.
This is currently not really an issue, but will be one, if the user may
supply valgrind flags.
Therefore a second thread is used to decouple the subcommand execution
and the reading from the TCP port.