Allow for multiple stack traces
This commit is a quick and dirty solution to #74 and similar issues. The
valgrind output looks something like this:

    <what>Invalid read of size 4</what>
    <stack>
        ...
    </stack>
    <auxwhat>Address 0x4aa3b28 is 0 bytes after a block of size 24 alloc'd</auxwhat>
    <stack>
        ...
    </stack>

`serde` (rightfully) did not allow deserializing the same field twice
into a single stack trace. Therefore the solution is to just let it read
into a vector. The code using it now just always reads the first entry
in this vector, effectively discarding the additional information, which
is not good, but easiest for now.