Update dependencies
Due to the current MSRV of Rust 1.51 (rather old), one needs to use that
version for `cargo update`: `cargo +1.51 update`. Note, that this drops
the unmaintained and vulnerable `atty` at the cost of a downgrade of
`colored` (fortunately without necessary code changes). One problem is,
that the older cargo does not handle the MSRV correctly and selects a
"too new" version of `is-terminal`, hence a manual adjustment of that
dependency is necessary:

    `cargo +1.51 update -p is-terminal --precise 0.4.7`