Install the deploy SSH keys
The [latest release workflow] failed due to the branch protection rules,
which prevents changes to the default branch without a pull request. The
exception for deploy keys were not relevant, as there is no deploy key
used as of now. Therefore this commit uses a newly added deploy SSH key
(RSA).

[latest release workflow]: https://github.com/jfrimmel/cargo-valgrind/actions/runs/10905905182/job/30265898820