Detect stack overflows and report them properly
This commit checks for a specific valgrind output string occurring when
a potential stack overflow is detected. The output is "parsed" a bit, in
order to produces a nicer user experience. The output looks like this:

    $ cargo valgrind run --manifest-path tests/corpus/Cargo.toml --bin issue-13
       Error: looks like the program overflowed its stack
        Info: valgrind says:
              ==1937526== Stack overflow in thread #1: can't grow stack to 0x1ffe801000
              ==1937526== Stack overflow in thread #1: can't grow stack to 0x1ffe801000
              ==1937526== Stack overflow in thread #1: can't grow stack to 0x1ffe801000
              ==1937526==  If you believe this happened as a result of a stack
              ==1937526==  overflow in your program's main thread (unlikely but
              ==1937526==  possible), you can try to increase the size of the
              ==1937526==  main thread stack using the --main-stacksize= flag.
              ==1937526==  The main thread stack size used in this run was 8388608.

This clearly states the most likely cause of the error and thus helps a
user to search the error. The actual valgrind output is written there as
well, since it contains the help how to diagnose this more/mitigate a
potential false positive by creating a bigger stack.

Fixes #13.