Cleanup host-lookup in driver code
The old code was very messy for multiple reasons: the code worked almost
entirely with `Vec<u8>` (a.k.a. byte strings) instead of actual strings,
and thus needed to perform all string operations by hand. The search was
also linear (although the performance penalty is negligible in this
case). Another bad thing is, that the code always executed the fallback
path of querying `rustc -vV` even if the normal path would have worked
(which are the majority of cases), therefore the code needed a redesign.

This PR pulls the search over a specific command output out into its own
function and runs the primary path first and only runs the fallback path
if necessary. This makes the code much easier to read and under- stand.
The new function also works on strings, making its implementation much
simpler.