Always run clippy as part of the style CI job
In contrast to the older approach (removed in #59), this simply fails
the `style` CI job if there are clippy warnings or errors instead of
commenting it to the pull request. This has the downside, that new lints
introduced in the stable Rust compiler might break users committing new
code, but at least the action is clear in such a case. The upside com-
pared to the commenting approach is, that it is simple and will always
work (even for new contributors).

Fixes #60.