Add test for behavior if tool is killed mid-run
This is the codified version of #36, just with a SIGKILL instead of the
mentioned SIGING (or SIGTERM). It is unclear, if the tool could even be-
have sane with SIGKILL, but this is the only thing available in the Rust
standard library. The test currently fails and thus shows that the issue
is reproducible.