Periodically test repository with beta compiler
It can happen, that new compiler versions introduce new behavior to the
Rust standard library, which might impact this tool (see #111 for an ex-
ample of such an issue). Therefore this job builds the code with the
current beta compiler to detect potential issues before they reach the
stable compiler/standard library. The jobs runs once per week and will
create a GitHub issue in case of a failing build.