Test a plain default crate by `cargo new --lib`
This is the minimal crate, where `cargo valgrind` must never report any
issues. If that test fails, then there is a leak in the Rust standard
library, which is [reported to be allowed][leak-issue]:

> Based on the consensus in rust-lang/rust#133574, non-default full
> leakcheck is "convenience if we can, but no hard guarantees", [...].

This test is the earliest integration test by naming it `a...`, so that
this very error is caught first and with a clearer error message thanks
to the descriptive test name.

[leak-issue]: https://github.com/rust-lang/rust/issues/135608#issuecomment-2597205627