Automatically delete `vgcore.*` created by test
If a program crashes while running it in Valgrind, a `vgcore.<pid>`-file
might be created in the current working directory (if core dumping is
active on the running machine). In order to not clutter the main project
directory, a new drop guard is implemented by this commit. This type
will automatically remove (all) found `vgcore.*`-files, but crucially
not only the one created by this specific test, as this is not easily
detectable. One should add this drop-guard to each crashing test, as
else the files will be left over, if only that specific test is run
(e.g. due to test filtering).