Add suppressions for different Rust std versions
As reported in #111 and #116, at least Rust Rust 1.83 and the upcoming
Rust 1.86 versions have possible leaks in the standard library. Since
those reported lints are not actionable by the user, they should be
suppressed. This was requested in #107 as well.

Thankfully, `valgrind` has support for suppression files, that even can
be auto-generated with the `--gen-suppressions=yes`-flag on normal
valgrind invocations, e.g. [like so][comment]:
```shell
$ cargo new --lib some-tests
$ cd some-tests
$ cargo test --lib
$ valgrind --leak-check=full --gen-suppressions=yes target/debug/deps/some_tests-$HASH
```
Those generated traces are the full backtrace of the application, which
maybe need to be trimmed down to match certain functions, which will
leak (as it is done in the Rust 1.86 case).

This commit adds the suppression files necessary to support all Rust
versions from the MSRV up to the current beta compiler (1.86). This was
checked using the following script:
```bash
for rust in $(seq -f '1.%g' 58 85) beta; do
    echo $rust
    VALGRINDFLAGS='--suppressions=../../suppressions/rust-1.83 --suppressions=../../suppressions/rust-1.86' cargo +$rust valgrind test --manifest-path tests/default-new-project/Cargo.toml -q
done
```

[comment]: https://github.com/jfrimmel/cargo-valgrind/issues/116#issuecomment-2720792281