Allow passing a suppression file list to valgrind
This commit allows to pass a list of paths to the `valgrind::execute()`-
command, which is converted into arguments of `--suppressions=<path>`
for each path. This allows to supply pre-defined suppression files to
valgrind. Note, that the list is currently empty, since there is no way
to embed the standard library suppressions yet.