Compile list of suppressions into binary
Previously, the user needed to provide the suppression files on disk in
order to make them readable by `valgrind` (as it unfortunately can only
read them by file). This is unfortunate, as `cargo valgrind` should pro-
vide a list of standard library suppressions without the need for extern
resources. Therefore the files are embedded into the binary.

This is done by two co-working parts:
1.  the build script, which reads all files in the `suppressions`-folder
    and writes their contents to a constant slice `SUPPRESSIONS`.
2.  the runtime code, which crates a temporary directory and writes the
    suppression contents into on file each in that directory and making
    use of them for `--suppressions`-options.

The code is rather simple when the I/O errors can be disregarded (as it
is done at the moment). I've opted to panic if the temporary directory
cannot be created, as this should almost never happen.