Use single suppression string instead of str-slice
Previously, the build-script generated a slice of strings, which the
application needed to concatenate together at runtime. This is slow and
entirely unnecessary, as the code generator can simply output the format
required by the application.
Therefore this commit replaced the list of strings with a single long
string, that is embedded in the binary, removing the need for any run-
time modifications of the string (it can be written directly to the tem-
porary file).