Remove hashes in Valgrind suppressions
Apparently, the hashes in the mangled function names are different in
each build of Rust (e.g. Rust 1.86.0-beta.5 and 1.86.0-beta.7 produce
different ones, 1.86.0 (stable) has different ones again). Therefore the
hashes get replaced with a wildcard in this commit.
This makes the suppressions applicable for all Rust 1.86 betas as well
as 1.86.0 itself.