Add support for newer Valgrind XML output versions
This commit mainly adds new possible values for the `ProtocolVersion`,
where previously only version 4 was supported. Version 5 already exists
for some time, but it was not used by default by Valgrind, therefore
this did not pose problems.
However, a new Version 6 was introduced with latest (current) Valgrind
version 3.26, which is also the new default. Thankfully, both protocol
versions are largely compatible with version 4:
* [version 5] introduces file-descriptor tracking and related additional
  error output. Since `--track-fds` is not used, this is not an issue.
* [version 6] introduces error summaries in the XML output mirroring the
  normal console output. Since this tool does some statics itself, it
  is not necessary to parse them.

Both of those changes are purely additive and, since our XML parsing
discards extraneous elements by default, does not cause the parsing to
fail.

The tests with the "good" (leak-free) output are added as well. The pre-
existing test with version 4 is renamed to make it clear and assertions
are added to check for the version.

[version 5]: https://sourceware.org/git/?p=valgrind.git;a=blob_plain;f=docs/internals/xml-output-protocol5.txt;hb=48d64d0e6bb72220bb2557be4f427a57038dfbc6
[version 6]: https://sourceware.org/git/?p=valgrind.git;a=blob_plain;f=docs/internals/xml-output-protocol6.txt;hb=7786b075abef51ca3d84b9717915f04b32950b32