Catch signals in program under test (like aborts)
This commit checks for the signal state of the valgrind process to see,
if valgrind propagated a signal (SIGSEGV/SIGABRT/etc.) from the program
under test. This allows to see, if the program did exit due to an abort
or similar and print an helpful error message like these in such cases:
* info: no memory error was detected, but the program was terminated by signal 6
* info: the program was terminated by signal 6

Tests with a new program with `process::abort()` are added (one without
leaks, once with a simple integer leak).