Make new signal code to be a no-op on Windows
Windows fundamentally does not support the signals as it uses another
mechanism. Currently the code is still built on windows in CI, so we
still need to compile on Windows as of now (see #101 for details).
Therefore this commit extracts the usage of the Unix `ExitStatusExt` to
a `cfg`ed function with a no-op function used when not on Unix. This
allows to compile it on all platforms (although the runtime behavior is
different).