Add support for `VALGRINDFLAGS=--track-fds=yes`
This commit adds the parsing support for the `--track-fds`-option in
Valgrind (currently needs to be passed via the `VALGRINDFLAGS`-variable)
for tracking file descriptor errors. This just needs the extra variant
in the error enumeration, since the rest of the tool is written to re-
port any non-leak errors as well. Therefore the actual change is just
the addition of the variant and its handling in the match statements,
i.e. 3 lines in total.
The remaining commit just adds a test case, based on the `libc`-crate
doing raw Unix syscalls with a wrong file descriptor, showing that the
XML can be parsed successfully.