Replace deprecated `assert_cmd::Command::cargo_bin`
This commit tries to resolve the issue #145 by using the macro for
obtaining the cargo command. This is required in order to not have a
breaking change in future Rust versions.
Unfortunately, the currently used version of [`assert_cmd`] does not
provide the suggested fix, so a newer version (at least `2.1`) is re-
quired.
Unfortunately this in turn requires a Rust version of at least 1.74, so
this commit also includes a bump of the MSRV. Only the minimal change in
dependencies is done, so that the diff is smaller. With the higher MSRV,
a bump of several dependencies can be done in a future commit.

Fixes of #145.

[`assert_cmd`]: https://docs.rs/assert_cmd/2.0.5/assert_cmd/cargo/fn.cargo_bin.html